controller action athorization impl. wth policies

+++ app/Http/Controllers/AssetController.php
@@ -152,6 +152,9 @@ class AssetController extends Controller
      */
     public function destroy(Asset $asset)
     {
+        // TODO: use athorizeResource() here, see UserController::__construct()
+        $this->authorize('delete', $asset);
+
         $asset->delete();
 
         return redirect()
